Move Navigation header into navigation dir. Added speed to gpx.
