+++ libosmscout/include/osmscout/navigation/Navigation.h
